+++ test/test_types.sh
@@ -45,7 +45,7 @@ newtest "print syntax, check first and last"
 expectpart "$(echo "" | $cligen_file -f $fspec -p)" 0  'a0 <v:ipv4addr>, callback();' 'r0 <v:string regexp:"\[a-z\]\[0-9\]\*">, callback();' 't0 <v:time>, callback();'
 
 newtest "dump syntax"
-expectpart "$(echo "" | $cligen_file -f $fspec -d)" 0  "pt types" "co a0"
+expectpart "$(echo "" | $cligen_file -f $fspec -d)" 0 "co a0"
 
 # regexps
 for x in 0 1 2 3 4 5; do